requirements.txt
vk_api>=11.9.9,<12.0.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.31.0,<3.0.0
